fix: apply clippy auto-fixes and fix E0716 in safety gateway
Ran cargo clippy --fix across all workspace crates, reducing warnings
from 249 to 0. Fixes include needless_range_loop, unnecessary_cast,
clone_on_copy, collapsible_if, needless_borrow, and unused imports.
Also fixes a temporary-value-dropped-while-borrowed error in
safety/gateway.rs that prevented symthaea from compiling.

+++ symthaea/crates/symthaea-stt/src/audio.rs
@@ -697,7 +697,7 @@ impl AudioProjector {
             };
 
             // Spread rotations across full 2048-bit space for maximum discrimination
-            let rotation = (level as i32 - 3) * 256;  // Center around 0
+            let rotation = (level - 3) * 256;  // Center around 0
             let rotated = basis.rotate(rotation);
             for w in 0..16 {
                 hv.words[w] ^= rotated.words[w];
